requirements.txt
# Core dependencies
numpy>=1.21.0
torch>=1.9.0
gymnasium>=0.28.1

# Visualization
matplotlib>=3.4.3
tensorboard>=2.6.0
pandas>=1.3.3

# Utils
tqdm>=4.62.3
pyyaml>=6.0

# Testing
pytest>=6.2.5
pytest-cov>=2.12.1

# Additional dependencies
seaborn>=0.11.0
pygame>=2.0.0
pettingzoo>=1.14.0
stable-baselines3>=1.3.0
networkx>=2.6.0
scipy>=1.7.0
sphinx>=4.2.0
pillow>=8.3.0
ipython>=7.30.0
jupyterlab>=3.2.0 